moving rake task to bulk import ideas engine + offences
