+++ back/.rubocop_todo.yml
@@ -399,6 +399,7 @@ Security/Open:
     - 'app/models/common_password.rb'
     - 'app/services/profanity_service.rb'
     - 'engines/free/email_campaigns/app/services/email_campaigns/tasks_service.rb'
+    - 'engines/commercial/bulk_import_ideas/lib/tasks/bulk_import_ideas_tasks.rake'
 
 # Offense count: 2
 # This cop supports safe auto-correction (--auto-correct).
